refactor: SlidingPiece 관련 패키징
