+++ src/main/java/dto/ChessBoardDTO.java
@@ -3,7 +3,13 @@ package dto;
 import domain.File;
 import domain.Rank;
 import domain.Square;
-import domain.piece.*;
+import domain.piece.King;
+import domain.piece.Knight;
+import domain.piece.Pawn;
+import domain.piece.Piece;
+import domain.piece.sliding.Bishop;
+import domain.piece.sliding.Queen;
+import domain.piece.sliding.Rook;
 
 import java.util.ArrayList;
 import java.util.List;
