+++ src/test/java/domain/piece/sliding/BishopTest.java
@@ -1,4 +1,4 @@
-package domain.piece;
+package domain.piece.sliding;
 
 import domain.*;
 import org.junit.jupiter.api.DisplayName;
@@ -10,15 +10,17 @@ import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class QueenTest {
-    @DisplayName("퀸은 상하좌우 대각선으로 여러칸 움직일 수 있다.")
+class BishopTest {
+    @DisplayName("비숍은 대각선으로 여러칸 움직일 수 있다.")
     @ParameterizedTest
     @MethodSource(value = "canMoveArguments")
     void canMove(final Square target) {
-        final Queen queen = new Queen(Team.BLACK);
+        // given
+        final Bishop bishop = new Bishop(Team.BLACK);
         final Square source = new Square(File.D, Rank.FOUR);
+
         // when
-        final boolean canMove = queen.canMove(source, target, new ChessBoard().getPieces());
+        final boolean canMove = bishop.canMove(source, target, new ChessBoard().getPieces());
 
         // then
         assertThat(canMove).isTrue();
@@ -26,24 +28,22 @@ class QueenTest {
 
     static Stream<Arguments> canMoveArguments() {
         return Stream.of(
-                Arguments.of(new Square(File.A, Rank.FOUR)),
-                Arguments.of(new Square(File.H, Rank.FOUR)),
-                Arguments.of(new Square(File.D, Rank.ONE)),
-                Arguments.of(new Square(File.D, Rank.EIGHT)),
                 Arguments.of(new Square(File.H, Rank.EIGHT)),
                 Arguments.of(new Square(File.A, Rank.ONE)),
                 Arguments.of(new Square(File.A, Rank.SEVEN)),
                 Arguments.of(new Square(File.G, Rank.ONE)));
     }
 
-    @DisplayName("퀸은 상하좌우 대각선을 제외하고 움직일 수 없다.")
+    @DisplayName("비숍은 상하좌우로 이동할 수 없다.")
     @ParameterizedTest
-    @MethodSource(value = "canNotMoveArguments")
+    @MethodSource("canNotMoveArguments")
     void canNotMove(final Square target) {
-        final Queen queen = new Queen(Team.BLACK);
+        // given
+        final Bishop bishop = new Bishop(Team.BLACK);
         final Square source = new Square(File.D, Rank.FOUR);
+
         // when
-        final boolean canMove = queen.canMove(source, target, new ChessBoard().getPieces());
+        final boolean canMove = bishop.canMove(source, target, new ChessBoard().getPieces());
 
         // then
         assertThat(canMove).isFalse();
@@ -51,8 +51,9 @@ class QueenTest {
 
     static Stream<Arguments> canNotMoveArguments() {
         return Stream.of(
-                Arguments.of(new Square(File.G, Rank.TWO)),
-                Arguments.of(new Square(File.G, Rank.THREE)),
-                Arguments.of(new Square(File.G, Rank.FIVE)));
+                Arguments.of(new Square(File.D, Rank.EIGHT)),
+                Arguments.of(new Square(File.D, Rank.ONE)),
+                Arguments.of(new Square(File.G, Rank.FOUR)),
+                Arguments.of(new Square(File.A, Rank.FOUR)));
     }
 }
